Move 'useUpcomingFiscalYears' to common hooks folder
